requirements.txt
pandas
pyyaml
pydub
sounddevice
numpy
mne